Remove user app to blwebapp
