+++ blwebapp/templates/blwebapp/example.js
