--- user/templates/user/login.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-{% load staticfiles %}
-
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-    <script src="{% static 'user/user.js' %}"></script>
-    <link href="{% static 'user/user.css' %}" rel="stylesheet">
-</head>
-<body>
-
-</body>
-</html>
